feat(be): consider userId on reservation

+++ apps/backend/src/reservation/reservation.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, UnprocessableEntityException } from '@nestjs/common'
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+  UnprocessableEntityException,
+} from '@nestjs/common'
 import { CreateReservationDto } from './dto/create-reservation.dto'
 import { PrismaService } from 'prisma/prisma.service'
 import { EventEmitter2 } from '@nestjs/event-emitter'
@@ -10,7 +15,10 @@ export class ReservationService {
     private readonly eventEmitter: EventEmitter2,
   ) {}
 
-  async createReservation(createReservationDto: CreateReservationDto) {
+  async createReservation(
+    createReservationDto: CreateReservationDto,
+    userId: number,
+  ) {
     const { menuIds, timeSlotId, ...rest } = createReservationDto
 
     const timeSlot = await this.prisma.timeSlot.findUnique({
@@ -27,6 +35,7 @@ export class ReservationService {
     const processedReservation = await this.prisma.reservation.create({
       data: {
         ...rest,
+        userId,
         timeSlotId,
         ...(menuIds?.length
           ? { menus: { connect: menuIds.map((id) => ({ id })) } }
@@ -62,20 +71,34 @@ export class ReservationService {
     })
   }
 
-  async getStoreReservations(storeId: number) {
+  async getStoreReservations(storeId: number, userId: number) {
+    const isStaff = await this.prisma.storeStaff.findUnique({
+      where: { userId_storeId: { userId, storeId } },
+      select: {
+        role: true,
+      },
+    })
+
+    if (!isStaff) {
+      throw new ForbiddenException('You are not a staff of this store')
+    }
+
     return await this.prisma.reservation.findMany({
       where: { storeId },
       include: {
         menus: true,
         timeSlot: true,
+        store: {
+          select: {},
+        },
       },
       orderBy: { id: 'desc' },
     })
   }
 
-  async getReservation(id: number) {
-    return await this.prisma.reservation.findUnique({
-      where: { id },
+  async getReservation(id: number, userId: number) {
+    return await this.prisma.reservation.findFirstOrThrow({
+      where: { id, userId },
       include: {
         menus: true,
         user: true,
@@ -85,7 +108,22 @@ export class ReservationService {
     })
   }
 
-  async removeReservation(id: number) {
+  async removeReservation(id: number, userId: number) {
+    const reservation = await this.prisma.reservation.findUnique({
+      where: { id },
+      select: { userId: true },
+    })
+
+    if (!reservation) {
+      throw new NotFoundException('Reservation not found')
+    }
+
+    if (reservation.userId !== userId) {
+      throw new ForbiddenException(
+        'You are not allowed to remove this reservation',
+      )
+    }
+
     const deletedReservation = await this.prisma.reservation.delete({
       where: { id },
     })
@@ -100,7 +138,19 @@ export class ReservationService {
     return deletedReservation
   }
 
-  async callReservation(id: number) {
+  async callReservation(id: number, userId: number) {
+    const isStaff = await this.prisma.reservation.findFirst({
+      where: {
+        id,
+        store: { staffs: { some: { userId } } },
+      },
+      select: { id: true },
+    })
+
+    if (!isStaff) {
+      throw new ForbiddenException('You are not a staff of this store')
+    }
+
     const reservation = await this.prisma.reservation.update({
       where: { id },
       data: {
